move craft and warehousecmd functions to their respective files
