+++ commands/warehouse.py
@@ -1,7 +1,7 @@
-from .decorators import send_typing_action
-from .decorators import log
 from brains import warehouse_crafting
 from utils import send
+from .decorators import log
+from .decorators import send_typing_action
 
 
 @send_typing_action
@@ -11,3 +11,8 @@ def warehouse(update, context):
     responses = warehouse_crafting(context)
     for response in responses:
         send(response, update, context)
+
+
+def warehousecmd(update, context):
+    context.args = update.effective_message.text.split('_')[1:]
+    warehouse(update, context)
